fix(router): fix of deactivation guard breaks the routing history
Since we can’t determine whether the user actually meant the `back` or
the `forward` using the popstate event.
We could use this as the solution for the issue of getting the url back
to it’s initial state,
right before `popstate` is triggered and when the guard returns `false`.
So by using the `History#go` method,
we can bring the user back to the page where he is supposed to be after
performing the action.
Add the `ngRouterPageId` in the `RestoredState`.
Add a solution to the navigation with `skipLocationChange`.
add a solution to the navigation with `replaceUrl`.
Add a solution to the navigation with `eager` `urlUpdateStrategy`
Add `cancelNavigationTransition` method to the router.

fixes #13586