fix(router): fix of deactivation guard breaks the routing history
Since we can’t determine whether the user actually meant the “back” or
the “forward” using the popstate event.
We could use this as the solution for the issue of getting the url back
to it’s initial state,
right before popstate is triggered and when the guard returns “false”.
So by using the `goTo` method,
we can bring the user back to the page where he is supposed to be after
performing the action.
add the pageId field and use it instead of navigationId.

fixes #13586