fix(router): change `currentPageId` to `ngRouterPageId`
Change the  `currentPageId` to `ngRouterPageId`.
Change the semicolon to a comma in the definition of
`NavigationTransition`, and update the logic.
Add a solution to the navigation with `skipLocationChange`.
Accept new golden file.
Add more testing arround restoredState
We change the state type's of setBrowserUrl to RestoredState
Replace state for the url in history when navigation was canceled by empty resolve